factor out identical code for handling final scheduled basal
