+++ lib/insulet/insuletSimulator.js
@@ -21,6 +21,7 @@ var util = require('util');
 
 var annotate = require('../eventAnnotations');
 var common = require('./common');
+var simulations = require('../commonSimulations');
 
 var twentyFourHours = 24 * 60 * 60 * 1000;
 
@@ -256,34 +257,7 @@ exports.make = function(config){
           }
         }
         else if (settings != null) {
-          var millisInDay = sundial.getMsFromMidnight(currBasal.time, currBasal.timezoneOffset);
-          var basalSched = settings.basalSchedules[currBasal.scheduleName];
-          if (basalSched == null || basalSched.length === 0) {
-            if (!currBasal.isAssigned('duration')) {
-              currBasal.duration = 0;
-              annotate.annotateEvent(currBasal, 'basal/unknown-duration');
-              currBasal = currBasal.done();
-            }
-          }
-          else {
-            for (var i = basalSched.length - 1; i >= 0; --i) {
-              if (basalSched[i].start <= millisInDay) {
-                break;
-              }
-            }
-            if (basalSched[i].rate === currBasal.rate) {
-              currBasal.duration = (i + 1 === basalSched.length ? twentyFourHours - millisInDay : basalSched[i + 1].start - millisInDay);
-              currBasal = currBasal.done();
-            }
-            else {
-              if (!currBasal.isAssigned('duration')) {
-                currBasal.duration = 0;
-                annotate.annotateEvent(currBasal, 'insulet/basal/off-schedule-rate');
-                annotate.annotateEvent(currBasal, 'basal/unknown-duration');
-                currBasal = currBasal.done();
-              }
-            }
-          }
+          currBasal = simulations.finalScheduledBasal(currBasal, settings, 'insulet');
         }
         else {
           currBasal.with_duration(0);
