+++ lib/tandem/tandemSimulator.js
@@ -20,6 +20,7 @@ var sundial = require('sundial');
 var util = require('util');
 
 var annotate = require('../eventAnnotations');
+var common = require('../commonSimulations');
 
 /**
  * Creates a new "simulator" for Tandem insulin pump data. The simulator has methods for events like
@@ -107,34 +108,7 @@ exports.make = function(config){
           }
         }
         else if (config.settings != null) {
-          var millisInDay = sundial.getMsFromMidnight(currBasal.time, currBasal.timezoneOffset);
-          var basalSched = config.settings.basalSchedules[currBasal.scheduleName];
-          if (basalSched == null || basalSched.length === 0) {
-            if (!currBasal.isAssigned('duration')) {
-              currBasal.duration = 0;
-              annotate.annotateEvent(currBasal, 'basal/unknown-duration');
-              currBasal = currBasal.done();
-            }
-          }
-          else {
-            for (var i = basalSched.length - 1; i >= 0; --i) {
-              if (basalSched[i].start <= millisInDay) {
-                break;
-              }
-            }
-            if (basalSched[i].rate === currBasal.rate) {
-              currBasal.duration = (i + 1 === basalSched.length ? 864e5 - millisInDay : basalSched[i + 1].start - millisInDay);
-              currBasal = currBasal.done();
-            }
-            else {
-              if (!currBasal.isAssigned('duration')) {
-                currBasal.duration = 0;
-                annotate.annotateEvent(currBasal, 'tandem/basal/off-schedule-rate');
-                annotate.annotateEvent(currBasal, 'basal/unknown-duration');
-                currBasal = currBasal.done();
-              }
-            }
-          }
+          currBasal = common.finalScheduledBasal(currBasal, config.settings, 'tandem');
         }
         else {
           currBasal.with_duration(0);
